refactor: centralize firebase module
